requirements-lambda+cdk+sec-frozen.txt
aenum==3.1.11
aiohttp==3.8.1
aiosignal==1.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
aws-cdk-lib==2.41.0
aws-cdk.aws-glue-alpha==2.41.0a0
awswrangler==2.16.1
backoff==2.1.2
bandit==1.7.4
beautifulsoup4==4.11.1
boto3==1.24.66
botocore==1.27.66
CacheControl==0.12.11
cattrs==22.1.0
cdk-nag==2.18.25
certifi==2022.6.15
charset-normalizer==2.0.12
commonmark==0.9.1
constructs==10.1.104
curlify==2.2.1
cyclonedx-python-lib==3.1.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
facebook-business==14.0.0
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
gremlinpython==3.6.1
html5lib==1.1
idna==3.3
isodate==0.6.1
jmespath==1.0.1
jsii==1.69.0
jsonpath-ng==1.5.3
lockfile==0.12.2
lxml==4.9.1
msgpack==1.0.4
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.2
openpyxl==3.0.10
opensearch-py==1.1.0
packageurl-python==0.10.3
packaging==21.3
pandas==1.4.4
pbr==5.10.0
pg8000==1.29.1
pip-api==0.0.30
pip-requirements-parser==31.2.0
pip_audit==2.4.4
ply==3.11
progressbar2==4.0.0
publication==0.0.3
pyarrow==7.0.0
pycountry==22.3.5
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
PyYAML==6.0
redshift-connector==2.0.908
requests==2.28.0
requests-aws4auth==1.1.2
resolvelib==0.8.1
rich==12.6.0
s3transfer==0.6.0
scramp==1.4.1
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stevedore==4.0.0
toml==0.10.2
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
yarl==1.8.1
